requirements.txt
django
requests
djangorestframework==3.15.2
